+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0 (2020-06-03)
+
+* Far faster implementation of `stats.coclustering`
+* Introducing `helpers.make_slurm_command`
+* Global submodule import (no longer need to import submodules individually)
+
 ## 0.1.0 (2020-03-06)
 
 * First release on PyPI.